+++ packages/terra-functional-testing/src/terra-cli/wdio/index.js
@@ -47,6 +47,11 @@ const cli = {
         describe: 'A flag to disable the selenium docker service.',
         default: false,
       },
+      ignoreScreenshotMismatch: {
+        type: 'boolean',
+        describe: 'A flag to ignore screenshot mismatch.',
+        default: false,
+      },
       externalHost: {
         type: 'string',
         describe: 'The host address the testing environment is connected to.',